third_party/blink/web_tests/http/tests/devtools/sources/debugger/source-frame-inline-breakpoint-decorations-expected.txt

Checks that JavaScriptSourceFrame show inline breakpoints correctly


Running: testAddRemoveBreakpoint
Setting breakpoint
PROMISE FAILURE: undefined
